Moved ListInput and ObjectInput to misc
